requirements.txt
Flask
flask-cors
google-generativeai
requests
beautifulsoup4